Move FedAvg Collab example to hello world
