+++ examples/README.md
@@ -77,7 +77,7 @@ When you open a notebook, select the kernel `nvflare_example` using the dropdown
 |------------------------------------------------------------------------------|--------------|-----------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | [Notebook for Hello Examples](./hello-world/hello_world.ipynb)               | -            | Notebook for examples below.                                                                                                                                    |
 | [Hello NumPy](./hello-world/hello-numpy/README.md)             | Numpy        | Example using [FedAvg](https://nvflare.readthedocs.io/en/main/apidocs/nvflare.app_common.workflows.fedavg.html) with Recipe API.      |
-| [Hello NumPy Collab](./hello-world/hello-collab/README.md)     | Numpy        | Reimplements the Hello NumPy workflow with the Collab API and ordinary Python function calls.                                                                    |
+| [Hello FedAvg Collab](./hello-world/hello-collab/README.md)   | PyTorch      | Implements a custom FedAvg workflow with the Collab API and ordinary Python function calls.                                                                       |
 | [Hello Cross-Site Validation](./hello-world/hello-numpy-cross-val/README.md) | Numpy        | Example using [CrossSiteEval](https://nvflare.readthedocs.io/en/main/apidocs/nvflare.app_common.workflows.cross_site_eval.html) controller workflow, and example using previous results without training workflow. |
 | [Hello Cyclic Weight Transfer](./hello-world/hello-cyclic/README.md)         | PyTorch      | Example using [CyclicController](https://nvflare.readthedocs.io/en/main/apidocs/nvflare.app_common.workflows.cyclic_ctl.html) controller workflow to implement [Cyclic Weight Transfer](https://pubmed.ncbi.nlm.nih.gov/29617797/). |
 | [Hello JAX](./hello-world/hello-jax/README.md)                                | JAX          | Example using an MNIST classifier with [FedAvg](https://arxiv.org/abs/1602.05629) and JAX/Flax as the deep learning training framework. |
