+++ examples/advanced/collab/README.md
@@ -1,9 +1,8 @@
 # Advanced Collab API Examples
 
-These self-contained examples demonstrate custom workflows built with the core
-Collab programming model. For a minimal introduction that mirrors
-`hello-world/hello-numpy`, start with
-[Hello NumPy Collab](../../hello-world/hello-collab/README.md).
+These self-contained examples demonstrate advanced custom workflows built with
+the core Collab programming model. For a minimal introduction, start with
+[Hello FedAvg with the Collab API](../../hello-world/hello-collab/README.md).
 
 Every entry point builds a `CollabRecipe` and runs it directly with the standard
 `SimEnv`:
@@ -11,7 +10,6 @@ Every entry point builds a `CollabRecipe` and runs it directly with the standard
 ```bash
 cd examples/advanced   # makes the collab package available to module imports
 
-python -m collab.hello_fedavg.hello_fedavg
 python -m collab.simple_split_learning.simple_split_learning
 python -m collab.async_aggregation.async_aggregation
 python -m collab.swarm.swarm --num-clients 3
@@ -22,7 +20,6 @@ python -m collab.swarm.swarm --num-clients 3
 
 | Example | Demonstrates |
 |---|---|
-| `hello_fedavg` | The Collab API in one file: `@collab.main`, `@collab.publish`, `collab.clients.train(...)`, per-site config |
 | `simple_split_learning` | Split learning on MNIST with client-side images and bottom model, server-side labels and top model, and direct activation/gradient exchange |
 | `async_aggregation` | In-time aggregation with a response callback |
 | `swarm` | Decentralized swarm learning with client-to-client calls |
@@ -47,7 +44,6 @@ To use another deployment mode, execute the same recipe with `PocEnv` or
 `ProdEnv` from `nvflare.recipe`; Collab has no separate runner or environment
 abstraction.
 
-The NumPy examples run in a base installation; `hello_fedavg` needs PyTorch.
 `simple_split_learning` needs PyTorch and torchvision and downloads MNIST on
 its first run.
 `pt_cifar10` needs PyTorch and torchvision; follow its
