+++ examples/hello-world/README.md
@@ -114,6 +114,23 @@ pip install -r requirements.txt
 python job.py
 ```
 
+#### [Hello FedAvg with the Collab API](./hello-collab/)
+Implement a custom FedAvg workflow with ordinary Python function calls and a
+small synthetic PyTorch model.
+
+**What you'll learn:**
+- Defining a custom server workflow with `@collab.main`
+- Publishing client training code with `@collab.publish`
+- Passing PyTorch state dictionaries and scalar losses directly
+- Configuring client-specific local epochs
+
+**Run it:**
+```bash
+cd hello-collab
+pip install -r requirements.txt
+python job.py
+```
+
 #### [Hello PyTorch Lightning](./hello-lightning/)
 Train an image classifier using PyTorch Lightning with federated learning.
 
@@ -211,20 +228,6 @@ pip install -r requirements.txt
 python job.py
 ```
 
-#### [Hello NumPy Collab](./hello-collab/)
-Implement the same NumPy workflow with ordinary Python function calls.
-
-**What you'll learn:**
-- Defining a custom server workflow with `@collab.main`
-- Publishing client training code with `@collab.publish`
-- Passing NumPy arrays and tuples directly between server and clients
-
-**Run it:**
-```bash
-cd hello-collab
-python job.py
-```
-
 ### Framework Integration
 
 #### [Hello Flower](./hello-flower/)
@@ -254,8 +257,8 @@ hello-<framework>/
 └── requirements.txt  # Dependencies
 ```
 
-The self-contained [Hello NumPy Collab](./hello-collab/) example keeps its
-published client function, server workflow, recipe, and launcher in one file.
+The self-contained [Hello FedAvg Collab](./hello-collab/) example keeps its
+published client method, server workflow, recipe, and launcher in one file.
 
 ### Client Code (`client.py`)
