+++ examples/hello-world/hello-collab/requirements.txt
@@ -1,1 +1,1 @@
-nvflare~=2.9.0rc
+nvflare[PT]~=2.9.0rc
